test: add unit tests for truncate_prompt functionality
